Fixed circular import and renamed files
