+++ pyaptamer/utils/__init__.py
@@ -11,13 +11,13 @@ __all__ = [
     "hf_to_dataset",
 ]
 
+from pyaptamer.utils._fasta_to_aaseq import fasta_to_aaseq
+from pyaptamer.utils._hf_to_dataset import hf_to_dataset
 from pyaptamer.utils._rna import (
     dna2rna,
     encode_rna,
     generate_all_aptamer_triplets,
     rna2vec,
 )
-from pyaptamer.utils.fasta_to_aaseq import fasta_to_aaseq
-from pyaptamer.utils.hf_to_dataset import hf_to_dataset
 from pyaptamer.utils.pdb_to_struct import pdb_to_struct
 from pyaptamer.utils.struct_to_aaseq import struct_to_aaseq
